Update the layout of the first sheet
- Move generateTemplate to utils
- Make optionsMap and context optional for parseRichText
